責務分割でtype_body_lengthを整理
